backend/requirements-test.txt
pytest
httpx
fastapi
requests
python-dotenv
cachetools